backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
alembic==1.12.1

# LangChain and OpenAI dependencies
langchain==0.0.340
langchain-openai==0.0.2
openai==1.3.5

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment and configuration
python-dotenv==1.0.0

# HTTP and networking
httpx==0.25.2
requests==2.31.0

# Date and time handling
python-dateutil==2.8.2

# Logging and monitoring
structlog==23.2.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: PostgreSQL support (uncomment if using PostgreSQL)
# psycopg2-binary==2.9.9

# Optional: MySQL support (uncomment if using MySQL)
# PyMySQL==1.1.0